YOUR-WORKSPACE-v0.9/.claude/skills/notion-upload-lv1/requirements.txt
notion-client==2.7.0
python-dotenv==1.2.1
